+++ src/main/java/com/example/tripit/user/repository/UserRepository.java
@@ -1,7 +1,6 @@
 package com.example.tripit.user.repository;
 
 import com.example.tripit.user.entity.UserEntity;
-import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -16,13 +15,12 @@ public interface UserRepository extends JpaRepository<UserEntity, Integer>{
     Boolean existsByNickname(String nickname);
 
     //nickname를 받아 DB테이블에서 회원을 조회하는 메소드 작성
-
     UserEntity findByEmail(String email);
 
     @Query("SELECT u.userId FROM user u WHERE u.email = :email")
-    Integer findUserIdByEmail(@Param("email") String email);
+    long findUserIdByEmail(@Param("email") String email);
 
-    Optional<UserEntity> findById(Integer userId);
+    Optional<UserEntity> findById(Long userId);
 
     //UserEntity findByNickname(String nickname);
